Renaming: Negation->Inverter and Deadline->Timeout
