+++ docs/uml/TypeHierarchy.uxf
@@ -16,7 +16,7 @@
   <element>
     <id>UMLClass</id>
     <coordinates>
-      <x>490</x>
+      <x>590</x>
       <y>280</y>
       <w>100</w>
       <h>30</h>
@@ -27,8 +27,8 @@
   <element>
     <id>UMLClass</id>
     <coordinates>
-      <x>380</x>
-      <y>280</y>
+      <x>590</x>
+      <y>320</y>
       <w>100</w>
       <h>30</h>
     </coordinates>
@@ -38,8 +38,8 @@
   <element>
     <id>UMLClass</id>
     <coordinates>
-      <x>600</x>
-      <y>280</y>
+      <x>590</x>
+      <y>240</y>
       <w>120</w>
       <h>30</h>
     </coordinates>
@@ -49,41 +49,41 @@
   <element>
     <id>Relation</id>
     <coordinates>
-      <x>420</x>
+      <x>500</x>
       <y>200</y>
       <w>110</w>
-      <h>100</h>
+      <h>160</h>
     </coordinates>
     <panel_attributes>lt=&lt;&lt;-</panel_attributes>
-    <additional_attributes>90.0;10.0;10.0;80.0</additional_attributes>
+    <additional_attributes>10.0;10.0;10.0;140.0;90.0;140.0</additional_attributes>
   </element>
   <element>
     <id>Relation</id>
     <coordinates>
       <x>530</x>
       <y>200</y>
-      <w>30</w>
-      <h>100</h>
+      <w>80</w>
+      <h>110</h>
     </coordinates>
     <panel_attributes>lt=&lt;&lt;-</panel_attributes>
-    <additional_attributes>10.0;10.0;10.0;80.0</additional_attributes>
+    <additional_attributes>10.0;10.0;10.0;90.0;60.0;90.0</additional_attributes>
   </element>
   <element>
     <id>Relation</id>
     <coordinates>
       <x>560</x>
       <y>200</y>
-      <w>110</w>
-      <h>100</h>
+      <w>50</w>
+      <h>70</h>
     </coordinates>
     <panel_attributes>lt=&lt;&lt;-</panel_attributes>
-    <additional_attributes>10.0;10.0;90.0;80.0</additional_attributes>
+    <additional_attributes>10.0;10.0;10.0;50.0;30.0;50.0</additional_attributes>
   </element>
   <element>
     <id>UMLClass</id>
     <coordinates>
-      <x>340</x>
-      <y>380</y>
+      <x>700</x>
+      <y>420</y>
       <w>100</w>
       <h>30</h>
     </coordinates>
@@ -93,7 +93,7 @@
   <element>
     <id>UMLClass</id>
     <coordinates>
-      <x>450</x>
+      <x>700</x>
       <y>380</y>
       <w>120</w>
       <h>30</h>
@@ -104,30 +104,30 @@
   <element>
     <id>Relation</id>
     <coordinates>
-      <x>380</x>
-      <y>300</y>
-      <w>60</w>
-      <h>100</h>
+      <x>620</x>
+      <y>340</y>
+      <w>100</w>
+      <h>120</h>
     </coordinates>
     <panel_attributes>lt=&lt;&lt;-</panel_attributes>
-    <additional_attributes>40.0;10.0;10.0;80.0</additional_attributes>
+    <additional_attributes>10.0;10.0;10.0;100.0;80.0;100.0</additional_attributes>
   </element>
   <element>
     <id>Relation</id>
     <coordinates>
-      <x>430</x>
-      <y>300</y>
+      <x>640</x>
+      <y>340</y>
       <w>80</w>
-      <h>100</h>
+      <h>70</h>
     </coordinates>
     <panel_attributes>lt=&lt;&lt;-</panel_attributes>
-    <additional_attributes>10.0;10.0;60.0;80.0</additional_attributes>
+    <additional_attributes>10.0;10.0;10.0;50.0;60.0;50.0</additional_attributes>
   </element>
   <element>
     <id>UMLNote</id>
     <coordinates>
-      <x>590</x>
-      <y>330</y>
+      <x>460</x>
+      <y>360</y>
       <w>140</w>
       <h>90</h>
     </coordinates>
